% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAttribution.R
\docType{package}
\name{geneAttribution}
\alias{geneAttribution}
\alias{geneAttribution-package}
\title{geneAttribution: Identification of candidate genes associated with noncoding genetic variation}
\description{
Identification of the most likely gene or genes through which variation at a given genomic
locus in the human genome acts. The most basic functionality assumes that the closer gene is
to the input locus, the more likely the gene is to be causative. Additionally, any empirical
data that links genomic regions to genes (e.g. eQTL or genome conformation data) can be used
if it is supplied in UCSC .bed file format.
A typical workflow requires loading gene models and empirical data, then running ga() on the
locus of interest
}

